pyproject.toml
[tool.poetry]
name = "aarthik-ai"
version = "0.1.0"
description = "Financial Intelligence Chatbot with LangGraph and Chainlit"
authors = ["Rudra"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Core Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"

# LangChain & LangGraph
langchain = "^0.3.12"
langchain-openai = "^0.2.12"
langchain-anthropic = "^0.3.0"
langgraph = "^0.2.55"
langchain-community = "^0.3.12"
langsmith = "^0.2.5"

# Vector Store & Embeddings
pinecone-client = "^5.0.1"
sentence-transformers = "^3.3.1"

# Database
pymongo = "^4.10.1"
motor = "^3.6.0"
redis = "^5.2.0"

# API Clients
httpx = "^0.28.1"
openai = "^1.57.0"

# Frontend
chainlit = "^1.3.2"

# Utilities
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
tiktoken = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
